+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `external` data option for `badge` views
 - Fixed a crush on open actions popup in `struct` view when a value has too much nested structure
 - Uniformed a monospace font family across all views
+- Fixed highlighting issues in `source` view
 
 ## 1.0.0-beta.65 (18-04-2022)